backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ML and AI
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.1
bitsandbytes>=0.41.0
datasets>=2.16.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# RAG and Embeddings
langchain>=0.1.0
langchain-community>=0.0.13
chromadb>=0.4.22
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# PDF Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
pypdf>=3.17.4

# Google Gemini (Fallback)
google-generativeai>=0.3.2

# Utilities
python-dotenv>=1.0.1
httpx>=0.26.0
aiofiles>=23.2.1
tqdm>=4.66.1
numpy>=1.26.0
pandas>=2.2.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.4
black>=24.1.1
isort>=5.13.2
